requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
email-validator==2.0.0
Pillow==10.0.0
flask-bcrypt==1.0.1
stripe==6.5.0
python-dotenv==1.0.0
flask-migrate==4.0.4
gunicorn==21.2.0
mysqlclient==2.2.0
PyMySQL==1.1.0
cryptography==41.0.3